Add update modal flow
